+++ celescope/flv_trust4_split/mapping_annotation.py
@@ -1,29 +1,10 @@
-import pandas as pd
-import numpy as np 
-import subprocess
 import glob
 
-from celescope.flv_trust4_split.summarize import Summarize
-from celescope.tools import utils
-from celescope.tools.step import Step, s_common
-from celescope.flv_trust4_split.__init__ import TOOLS_DIR, CHAIN, PAIRED_CHAIN
-from celescope.flv_trust4_split import trust_utils as tr
-from celescope.tools.plotly_plot import Bar_plot
+import celescope.flv_trust4.mapping_annotation as tools_anno
+from celescope.flv_trust4.summarize import Summarize
 
 
-def run_mapping(rds, contig, sample, outdir, assign):
-    cmd = (
-        f'Rscript {TOOLS_DIR}/VDJmapping.R '
-        f'--rds {rds} '
-        f'--VDJ {contig} '
-        f'--sample {sample} '
-        f'--outdir {outdir} '
-        f'--assign_file {assign}'
-    )
-    subprocess.check_call(cmd, shell=True)
-
-
-class Mapping_annotation(Step):
+class Mapping_annotation(tools_anno.Mapping_annotation):
     """
     ## Features
 
@@ -41,7 +22,7 @@ class Mapping_annotation(Step):
 
     """
     def __init__(self, args, display_title=None):
-        Step.__init__(self, args, display_title=display_title)
+        super().__init__(args, display_title=display_title)
 
         self.seqtype = args.seqtype
         self.match_dir = args.match_dir
@@ -63,108 +44,11 @@ class Mapping_annotation(Step):
             self.Celltype = {'Plasma_cells','B_cells','Mature_B_cell','Plasma cells','B cells','Bcells'}
             self._name = "Bcells"
 
-    @staticmethod
-    def _parse_seqtype(seqtype):
-        return CHAIN[seqtype], PAIRED_CHAIN[seqtype]
-
-    @staticmethod
-    def parse_clonotype(outdir):
-        """Generate clonotypes table in html.
-        """
-        df_clonotypes=pd.read_csv(f'{outdir}/../03.summarize/clonotypes.csv', sep=',')
-        df_clonotypes['ClonotypeID'] = df_clonotypes['clonotype_id'].apply(lambda x: x.strip('clonetype'))
-        df_clonotypes['Frequency'] = df_clonotypes['frequency']
-        df_clonotypes['Proportion'] = df_clonotypes['proportion'].apply(lambda x: f'{round(x*100, 2)}%')
-        df_clonotypes['CDR3_aa'] = df_clonotypes['cdr3s_aa'].apply(lambda x: x.replace(';', '<br>'))
-        return df_clonotypes
-
-    @utils.add_log
-    def annotation_process(self):
-        """Generate metrics, clonotypes table in html.
-        """
-        df = pd.read_csv(self.contig_file)
-        annotation_summary = tr.get_vj_annot(df, self.chains, self.paired_groups)
-        for anno in annotation_summary:
-            self.add_metric(anno['name'], anno['value'], anno.get('total'), anno.get('help_info'))
-
-        df_clonotypes = self.parse_clonotype(self.outdir)
-        title = 'Clonetypes'
-        table_dict = self.get_table_dict(
-            title = title,
-            table_id = 'clonetypes',
-            df_table = df_clonotypes[['ClonotypeID', 'CDR3_aa', 'Frequency', 'Proportion']]
-        )
-        self.add_data(table_dict=table_dict)
-
-        df_clonotypes['ClonotypeID'] = df_clonotypes['ClonotypeID'].astype("int")
-        df_clonotypes.sort_values(by=['ClonotypeID'], inplace=True)
-        Barplot = Bar_plot(df_bar=df_clonotypes).get_plotly_div()
-        self.add_data(Barplot=Barplot)
-
-    @utils.add_log
-    def mapping_process(self):
-        """Mapping result with matched scRNA.
-        """
-        run_mapping(self.rds,self.contig,self.sample,self.outdir,self.assign_file)
-        meta = pd.read_csv(glob.glob(f'{self.outdir}/{self.sample}_meta.csv')[0])
-        metaTB = meta[meta['CellTypes'].isin(self.Celltype)]
-        mappedmeta = meta[meta['Class']=='T/BCR']
-        mappedmetaTB = mappedmeta[mappedmeta['CellTypes'].isin(self.Celltype)]
-        
-        Transcriptome_cell_number = meta.shape[0]
-        TB_cell_number = metaTB.shape[0]
-        Mapped_Transcriptome_cell_number = mappedmeta.shape[0]
-        Mapped_TB_cell_number = mappedmetaTB.shape[0]
-        mapping_summary=[]
-
-        mapping_summary.append({
-            'item': 'Cell Number in Matched transcriptome',
-            'count': Transcriptome_cell_number,
-            'total_count': np.nan
-        })
-        mapping_summary.append({
-            'item': 'Cell Number Successfully Mapped to transcriptome',
-            'count': Mapped_Transcriptome_cell_number,
-            'total_count': Transcriptome_cell_number
-        })
-        mapping_summary.append({
-            'item': f'{self._name} Number in Matched transcriptome',
-            'count': TB_cell_number,
-            'total_count': np.nan
-        })
-        mapping_summary.append({
-            'item': f'Cell Number Successfully Mapped to {self._name} in transcriptome',
-            'count': Mapped_TB_cell_number,
-            'total_count': TB_cell_number
-        })
-
-        stat_file = self.outdir + '/mapping.txt'
-        sum_df = pd.DataFrame(mapping_summary, columns=['item', 'count', 'total_count'])
-        utils.gen_stat(sum_df, stat_file) 
-    
-    def run(self):
-        self.annotation_process()
-        
-        try:
-            if self.rds and self.assign_file:
-                self.mapping_process()
-        except AttributeError:
-            print("rds file and type file do not exist" + "\n" )
-        except ZeroDivisionError:
-            print(f"Not found auto-assigned {self._name} in matched sc-RNA")
-
 
 def mapping_annotation(args):
     with Mapping_annotation(args, display_title="V(D)J Annotation") as runner:
         runner.run()
 
-    
+
 def get_opts_mapping_annotation(parser, sub_program):
-    parser.add_argument('--seqtype', help='TCR or BCR',
-                        choices=['TCR', 'BCR'], required=True)
-                        
-    if sub_program:
-        s_common(parser)
-        parser.add_argument('--match_dir', help='scRNA-seq match directory', required=True)
-        parser.add_argument('--contig_file', help='filtered contig annotation file', required=True)  
-    return parser
+    tools_anno.get_opts_mapping_annotation(parser, sub_program)
